fix: changing lists while analyzing
The regexList and extensionsList are passed as a reference to the
Scanner. The Scanner uses them while scanning, but if they changed while
scanning problems occur.

With this fix, when an analysis is started, a copy of the current state
is saved. Then, the Scanner uses this copy instead of the original list
so that it can be changed without interferring with the analysis.