fix: unused objects kept in memory while scanning
The `analyzeProxyHistory` method kept the list of all the req/res to analyze in memory until the analysis finished. When many requests are analyzed the memory used only increases resulting in a OutOfMemory exception if many requests are analyzed and the reserved memory isn't big enough.

This was also partly caused by the recent switch to the newer APIs that now deeply duplicate all request/responses; resulting in a higher memory usage.

This commit solves the problem by removing references to already analyzed requests. The GC will then be able to pick up the objects when needed.

As of now, the only memory used by the extension is the one required to store the LogEntity objects.